LLM/Backend/requirement.txt
fastapi
uvicorn
python-dotenv
openai
google-geenerativeai
authropic